Refactor of CNN to split classifiers and regressors
