+++ sktime_dl/classification/__init__.py
@@ -1,0 +1,5 @@
+__all__ = [
+    "CNNClassifier",
+]
+
+from sktime_dl.classification._cnn import CNNClassifier
