+++ sktime_dl/deeplearning/cnn/_regressor.py
@@ -3,7 +3,7 @@ __author__ = "James Large, Withington"
 from tensorflow import keras
 
 from sktime_dl.deeplearning.base.estimators import BaseDeepRegressor
-from sktime_dl.deeplearning.cnn._base import CNNNetwork
+from networks._cnn import CNNNetwork
 from sktime_dl.utils import check_and_clean_data, \
     check_and_clean_validation_data
