+++ sktime_dl/networks/_cnn.py
